requirements.txt
flask
python-dotenv
requests
pydub
uuid 